requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
SQLAlchemy==2.0.36
python-dotenv==1.0.1
requests==2.32.3
ldap3==2.9.1
